% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDtools_functions.r
\name{CompareGSDs}
\alias{CompareGSDs}
\title{Compare two grain size distribution samples to see if they are different}
\usage{
CompareGSDs(GSD1, GSD2, n1, n2, P = seq(5, 95, 5), alpha = 0.05)
}
\arguments{
\item{GSD1}{is a data frame containing grain sizes (in mm) in the first col-
umn, and the proportion of the distribution finer in the second column for
the first sample.}

\item{GSD2}{is a data frame containing grain sizes (in mm) in the first col-
umn, and the proportion of the distribution finer in the second column for
the second sample.}

\item{n1}{is  the total number of observations upon which the cumulative
frequency distribution in \code{GSD1} is based}

\item{n2}{is the total number of observations upon which the cumulative
frequency distribution in \code{GSD2} is based}

\item{P}{numeric vector of percentiles to be compared. The default is to
compare the 5th to the 95th percentile, in increments of 5}

\item{alpha}{the desired confidence level for which to calculate a
confidence interval. The default is to set alpha = 0.05, for a 95\% Confidence
Interval}
}
\description{
\code{CompareGSDs} is a function that takes two cumulative frequency dist-
ributions with the same format as the data frames produced by MakeCFD, and
uses an inverse transform approach to numerically determine whether per-
centile estimates from the two samples are statistically different. The
analysis requires only the cumulative frequency distribution for each sample
and the number of stones measured to generate the distribution.
}
\details{
The function returns a data frame listing the percentile being compared, the
estimate of that percentile for sample 1, the estimate for sample 2, and a
logical variable that indicates whether the percentiles are statistically
different or not.
}
