% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDtools_functions.r
\name{CompareRAWs}
\alias{CompareRAWs}
\title{Compare two sets of grain size measurements to see if they are different}
\usage{
CompareRAWs(OBS1, OBS2, P = seq(5, 95, 5), alpha = 0.05)
}
\arguments{
\item{OBS1}{is a vectory containing indivuala b axis measurments (in mm) for
the first sample.}

\item{OBS2}{is a vectory containing indivuala b axis measurments (in mm) for
the second sample.}

\item{P}{numeric vector of percentiles to be compared. The default is to
compare the 5th to the 95th percentile, in increments of 5}

\item{alpha}{the desired confidence level for which to calculate a
confidence interval. The default is to set alpha = 0.05, for a 95\% Confidence
Interval}
}
\description{
\code{CompareRAWs} is a function that takes two sets of grain size observations
and uses a resampling approach with replacement to estimate whether or not the
grain sizes for a percentile of interest from the two sample sets are statistically
different. The analysis requires the entire set of individual b-axis measurements.
}
\details{
The function returns a data frame listing the percentile being compared, the
estimate of that percentile for sample 1, the estimate for sample 2, and a
logical variable that indicates whether the percentiles are statistically
different or not.
}
