% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDtools_functions.r
\name{MakeCFD}
\alias{MakeCFD}
\title{Generate a data frame containing cumulative frequency distribution data}
\usage{
MakeCFD(obs, increment = 0.5, count = FALSE, plot = FALSE)
}
\arguments{
\item{obs}{vector containing b-axis measurements of gravel bed surface}

\item{increment}{an optional parameter to control the size of the grain size
classes used for the analysis (1.0 = 1.0 phi intervals, 0.5 = 1/2 phi
intervals, 0.25 = 1/4 phi intervals, etc.). Default value is 0.5.}

\item{count}{optional flag to record the number of observations falling in
each size class as an additional variable in the output data frame}

\item{plot}{optional flag to produce a graph of the resulting grain size
distribution}
}
\description{
\code{MakeCFD} is a function that creates a data frame containing the
cumulative grain size distribution information for a sample of n observations
of b axis diameter (in mm).
}
\details{
The function returns a data frame listing upper bounds of each size class
containing data, as well as the cumulative proportion of the observations
that fall below the given grain size. The output of this function is the
required input for other functions in this package. In order to retain
information about the sample size, there is the option to output the total
counts for each size class, as well.
}
