% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDtools_functions.r
\name{PolyCI}
\alias{PolyCI}
\title{Generate a polygon representing the confidence bounds for a grain size distribution}
\usage{
PolyCI(cfd, n, P = seq(1, 99, 1), equaltail = T, alpha = 0.05,
  plot = FALSE)
}
\arguments{
\item{cfd}{is a data frame providing a list grain sizes in the first variable
and the corresponding cumulative proportion finer in the second. The
grain sizes should be recorded in mm, and the proportion finer in [0,1].}

\item{n}{is the total number of observations upon which the cumulative
frequency distribution in \code{cfd} is based}

\item{P}{numeric vector of percentiles defining the polygon vertices,
with values in [0,100].}

\item{equaltail}{is a logical variable that determines whether the calculations
of the confidence interval are based on an approximation with equal areas in
each tail (the default), or based on the exact binomial solution with a coverage
of at least 95\%}

\item{alpha}{the desired confidence level for which to calculate a
confidence interval in [0,1].}

\item{plot}{optional flag to produce a graph of the resulting grain size
distribution}
}
\description{
\code{PolyCI} is a function that generates a data frame containing the
coordinates defining a polygon representing the confidence bounds around
a bed surface grain size distribution. The function contains an option to
generate a plot of the distribution showing the estimates of the percentiles
between the D5 and the D95, as well as the confidence limits about those
estimates.
}
\details{
The function returns a polygon object containing y coordinates that range
from 0 to 1 (i.e. the proportion finer), and corresponding grain sizes that
define the upper and lower bounds to the grain size confidence interval.
}
