% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDtools_functions.r
\name{QuantBD}
\alias{QuantBD}
\title{Calculate the confidence interval for a quantile 'p'}
\usage{
QuantBD(n, p, alpha = 0.05)
}
\arguments{
\item{n}{the sample size}

\item{p}{the desired quantile to be estimated from the sample in the range [0,1]}

\item{alpha}{the confidence level (default value is 0.05)}
}
\description{
\code{QuantBD} uses the  binomial distribution to compute a
near-symmetric distribution-free confidence interval for a quantile 'p'.
Returns indices for the order statistics, along with coverage probability.
}
\details{
The function returns the indices representing a percentile confidence interval
that contains the population quantile of interest, assuming a given
confidence level. The script also returns the probability coverage
for the confidence interval. Finally, the function returns an approximation
of the confidence interval that has equal areas in the tails of the distribution
}
